Make Django, httplib2, jinja2, docutils optional
- skip tests if optional dependencies not found
- added CI test with minimal requirements.
- setup.py not yet updated, since I anyway plan to move to pyproject.toml
- without Django, the ShelveRecordStore gets better tested; accordingly, added sorting of output records by timestamp to ShelveRecordStore

+++ test/unittests/test_recordstore.py
@@ -133,7 +133,7 @@ class MockProject(object):
     name = "TestProject"
 
 
-def setup():
+def setUpModule():
     global django_store1, django_store2, django_dir
     django_dir = tempfile.mkdtemp(prefix='sumatra-test-')
     cwd_before = os.getcwd()
@@ -145,12 +145,15 @@ def setup():
         django_store1 = django_store.DjangoRecordStore(db_file="test.db")
     if django_store2 is None:
         django_store2 = django_store.DjangoRecordStore(db_file="test2.db")
-    django_store.db_config.configure()
+    try:
+        django_store.db_config.configure()
+    except ImportError as err:
+        raise unittest.SkipTest(err)
     vcs_list.append(sys.modules[__name__])
     os.chdir(cwd_before)
 
 
-def teardown():
+def tearDownModule():
     global django_store1, django_store2, django_dir
     del sumatra.launch.MockLaunchMode
     del sumatra.datastore.MockDataStore
@@ -167,8 +170,10 @@ class BaseTestRecordStore(object):
         os.chdir(self.dir)
 
     def tearDown(self):
-        django_store1.delete_all()
-        django_store2.delete_all()
+        if django_store1:
+            django_store1.delete_all()
+        if django_store2:
+            django_store2.delete_all()
         os.chdir(self.cwd_before_test)
         shutil.rmtree(self.dir)
 
@@ -451,11 +456,15 @@ class TestHttpRecordStore(unittest.TestCase, BaseTestRecordStore):
 
     def __init__(self, *args, **kwargs):
         unittest.TestCase.__init__(self, *args, **kwargs)
-        self.real_httplib = http_store.httplib2
+        try:
+            self.real_httplib = http_store.httplib2
+        except AttributeError:
+            self.real_httplib = None
         http_store.httplib2 = MockHttpLib()
 
     def __del__(self):
-        http_store.httplib2 = self.real_httplib
+        if self.real_httplib:
+            http_store.httplib2 = self.real_httplib
 
     def setUp(self):
         BaseTestRecordStore.setUp(self)
@@ -544,6 +553,8 @@ class TestModuleFunctions(unittest.TestCase):
     def test_get_record_store_http(self, ):
         self.assertIsInstance(get_record_store("http://records.example.com/"),
                               http_store.HttpRecordStore)
+        self.assertIsInstance(get_record_store("http://testuser:abc123@0.0.0.0:55052/records/"),
+                              http_store.HttpRecordStore)
 
     def test_get_record_store_shelve(self):
         store = shelve_store.ShelveRecordStore(shelf_name="test_record_store.shelf")
@@ -562,6 +573,6 @@ class TestModuleFunctions(unittest.TestCase):
 
 
 if __name__ == '__main__':
-    setup()
+    setUpModule()
     unittest.main()
-    teardown()
+    tearDownModule()
